requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.2.0
boto3==1.20.26
botocore==1.23.26
certifi==2021.10.8
charset-normalizer==2.0.9
colorama==0.4.4
configparser==5.2.0
crayons==0.4.0
et-xmlfile==1.1.0
execnet==1.9.0
idna==3.3
importlib-metadata==4.8.3
iniconfig==1.1.1
jmespath==0.10.0
openpyxl==3.0.9
packaging==21.3
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
pytest-forked==1.4.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-xdist==2.5.0
python-dateutil==2.8.2
python-dotenv==0.19.2
requests==2.26.0
s3transfer==0.5.0
selenium==3.141.0
six==1.16.0
tinys3==0.1.12
toml==0.10.2
typing-extensions==4.0.1
urllib3==1.26.7
webdriver-manager==3.5.2
zipp==3.6.0
